requirements.txt
streamlit>=1.28.0
pyserial>=3.5
pandas>=2.0.0
plotly>=5.17.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
firebase-admin>=6.2.0
requests>=2.31.0
